Add nova venda.

+++ Teste_de_conhecimento_Campos_Dealer/Models/AddVendaViewModel.cs
@@ -9,6 +9,6 @@ namespace Teste_de_conhecimento_Campos_Dealer.Models
         public int qtdVenda { get; set; }
         public int vlrUnitarioVenda { get; set; }
         public DateTime dathVenda { get; set; }
-        public float vlrTotalVenda { get; set; }
+        public int vlrTotalVenda { get; set; }
     }
 }
